% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{download_data}
\alias{download_data}
\title{Download Car Cards}
\usage{
download_data(gsheet_url)
}
\arguments{
\item{gsheet_url}{A string with gsheet URL with Car Cards}
}
\value{
A data frame with Car Cards
}
\description{
Download Car Cards
}
