% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{download_data_sql}
\alias{download_data_sql}
\title{Download Car Cards from mySQL server}
\usage{
download_data_sql(dbname, host, username, password)
}
\arguments{
\item{dbname}{String with database name}

\item{host}{String with database host}

\item{username}{String with database user}

\item{password}{String with database password}
}
\value{
Content of \code{cards} SQL table
}
\description{
Download Car Cards from mySQL server
}
