% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{draw_starting_player}
\alias{draw_starting_player}
\title{Determine which player will start the game}
\usage{
draw_starting_player(players_n)
}
\arguments{
\item{players_n}{Number of players}
}
\value{
Starting player number
}
\description{
Determine which player will start the game
}
