% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{get_card_image}
\alias{get_card_image}
\title{Compile card image}
\usage{
get_card_image(card)
}
\arguments{
\item{card}{1-row data frame with card to be visualized}
}
\value{
Compiled card image
}
\description{
Compile card image
}
