% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{split_cards}
\alias{split_cards}
\title{Split cards between players an shuffle}
\usage{
split_cards(cards)
}
\arguments{
\item{cards}{A data frame with Car Cards}
}
\value{
A list of data frames with each player's Car Cards
}
\description{
Split cards between players an shuffle
}
