% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{upload_data_sql}
\alias{upload_data_sql}
\title{Upload processed Car Cards to mySQL server}
\usage{
upload_data_sql(cards, dbname, host, username, password, newdb = FALSE)
}
\arguments{
\item{cards}{A data frame with Car Cards}

\item{dbname}{String with database name}

\item{host}{String with database host}

\item{username}{String with database user}

\item{password}{String with database password}

\item{newdb}{Logical indicating if new \code{cards} table should be created}
}
\value{
Content of \code{cards} SQL table
}
\description{
Upload processed Car Cards to mySQL server
}
